backend/requirements.txt
Flask
tensorflow
numpy
Pillow
opencv-python
flask-cors